Remove titlebar and include side buttons in quick chat
This adds buttons to the right of the chat box. These buttons were previously on the quick pick's titlebar, but in order to have a smooth transition between Command Palette and Quick Chat we wanted to move away from the title bar.
